iio: light: cm32181: Add workaround for ACPI tables which list SMBus ARA address
On some ACPI platforms the ACPI tables contain I2cSerialBusV2 resources
for both the SMBus alert ARA address (0x0c), which is used for reading
interrupt status when using SMBus alerts, as well as a second
I2cSerialBusV2 resource with the actual I2C address of the CM3218 chip.

This commit adds a workaround for this, which when this is the case
creates another i2c-client for the second I2cSerialBusV2 resource and
returns -ENODEV.

As this new i2c-client has the same ACPI modalias our probe function
will get called a second time, for the new client, which does have
the correct address and we continue with binding against that client.

This has been tested on an Asus T100TA.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>